Fix Timeout Utils Memory Leak (#24116)
By reusing the same deferred promise, it was leading to promise chains
connected to the promise which would only be garbage collected on
promise failure, which only happens rarely as it leads to test failures.
This was causing gigabytes of memory to be leaked as those chained
promise and their references were essentially pinned in memory. This was
leading to out memory issues in the most degenerate case like when
running a test in a loop to reproduce failure.