Rollback doesn't work if you clear the batch after creating the checkpoint (#24128)
Rollback requires that the batch will remain intact between the
checkpoint and when rollback is called. This PR adds a test case to
illustrate this, which tries popping the current batch (clearing it) in
that window.

Note: `rollback` is intended for use only by ContainerRuntime's
`orderSequentially`, and CR already blocks flush (which is how you pop a
batch).