Make FieldSchema compatible with SimpleFieldSchema (#24123)
## Description

This is a start of a larger project to implement the SimpleSchema
interfaces with the actual public schema types.

This aligns FieldSchema with SimpleFieldSchema, but keep the actual
interface implementation alpha for now.
Next will be to align more schema types in this way with alpha
interfaces.

Once done for all simple schema types, ITreeAlpha can become alpha, and
these interfaces can begin stabilization.

## Breaking Changes

Users of the internal SimpleSchema APIs have had fields renamed to avoid
collisions.