fix(ci): Fix pipeline that calculates bundle size baselines (#24125)
## Description

Fixes the pipeline in the public ADO project that calculates bundle size
baselines, which
[broke](https://dev.azure.com/fluidframework/public/_build?definitionId=48&_a=summary)
when we merged https://github.com/microsoft/FluidFramework/pull/24044.
The root cause is that a test introduced in that PR depends on webpack
having run on the bundle-size-tests package to produce a bundle, which
the test then checks. The main `build - client pipeline` happens to run
webpack after the build script, but the pipeline that calculates bundle
size baselines does not, and thus it started failing.

The way this PR fixes the pipeline is by making it so it *does not run
the repo tests*, since it doesn't have any reason to. That's already
taken care of by `build - client packages`. That also saves us a lot of
runtime for it:


![image](https://github.com/user-attachments/assets/11e5c26d-7b30-4475-94d1-1af7874f9017)

### Side discussion

The PR that introduced the test tried to ensure webpack ran [by
declaring fluid-build
dependencies](https://github.com/microsoft/FluidFramework/blob/529fae8207ab32e113b8d5b45f0ade4dd01f0313/examples/utils/bundle-size-tests/package.json#L85-L90);
but we don't seem to use fluid-build to run tests anywhere, so this
didn't have any effect. Do we need to update our test npm scripts so
they use fluid-build, like our build ones do? It's not entirely clear to
me that we can easily do that.