LocalServerStress: Reenable skipped tests (#24132)
I believe the non-deterministic failures seen by a few seeds in this
test was fixed by recent improvement to the synchronization logic, that
was [previously
buggy](https://github.com/microsoft/FluidFramework/pull/24105/files#r2006610929).
Failed synchronization would account for the behavior seen in the test.
Before the update it was possible reproduce the issue in the gate and
locally by running repeated iterations of the test. I found it would hit
locally within about 150 iterations. I've run about 5000 iterations
locally, and 400 in the gate, and have not seen the issue reproduce


[AB#33713](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/33713)