ContainerRuntime: Use RunCounter for batches and reentrancy (#24134)
Adds a new class called RunCounter which simply counts the number of
nested runners it is used over. This pattern is used multiple times in
the container runtime, and encapsulating it allows us to remove a number
of error prone mutable variable from the ContainerRuntime and replace
them with immutable variables. It also reduces some coupling between the
different classes of batching, orderSequentially, and resubmit, and
makes them explicitly use the batchRunner, rather than depending on side
effects of orderSequentially. This will also allow us to further evolve
orderSequentially for things like StagingMode.

---------

Co-authored-by: Mark Fields <markfields@users.noreply.github.com>