Swap order of client comparison in fuzz harnesses (#23925)
If validation functions assert equality in the same order as the
arguments they are given, this causes assertion errors of the form
"actual: foo, expected: bar" to read as though the readonly client's
value is the expected one. Though this does not *always* give the
correct fix for eventual consistency issues, it oftentimes is and is
usually the right way to think about it.

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>