refactor(merge-tree): Clean up Perspective API (#24104)
## Description

This cleans up the perspective API by moving responsibility of
`nextSegment` / `prevSegment` to `MergeTree` and changing the typing to
minimally expose other concepts. The motivation for this is I would like
to propagate some of the API usage of `Perspective`s into `SharedString`
(e.g. via `MergeTreeTextHelper`) as an `@internal` package export, and I
want to avoid package exporting a bunch of concepts we previously took
effort to hide.

With these changes, we should be able to export the `Perspective`
interface plus some ways to get common ones and use them in
`SharedString`.

Co-authored-by: Abram Sanderson <absander@microsoft.com>