build(experimental-tree): Update doc model config to not conflict with production tree package (#24141)
We have two packages with the unscoped package name "tree" -
`@fluidframework/tree` and `@fluid-experimental/tree`. The API doc model
artifacts generated by API-Extractor are, by default, named for the
unscoped portion of the package name. As a part of our build, we collect
all of those API model artifacts and drop them into a single flat
directory to be uploaded as a build artifact for later consumption by
our website build. As things were previously configured, the
experimental tree package's artifact was clobbering our public tree
package's. This was resulting in our public tree package's API
documentation not being available to our website build.

This PR fixes the issue by renaming the experimental tree package's
artifact to `experimental-tree` instead of `tree`. This matches the
existing workaround we had in place to disambiguate the API report
generated by that package.