Align object schema with simple object node schema, and expose as alpha (#24143)
## Description

Several tweaks to the typing of object node schema have been made to
allow exposing an `@alpha` `ObjectNodeSchema` type.

See changeset for details.

This is part of a larger effort to align the various TreeNodeSchema with
the SimpleNodeSchema interfaces to provide friendly ways to inspect
schema information.

## Breaking Changes

The typing tweaks are not expected to break any existing valid use, bit
it is possible some exceptionally picky code could be impacted. See the
changeset for exact details.