(tree) In TestTreeProviderLite, provide capability to pause and resume processing of messages (#24111)
This change adds the capaibility to pause and resume inbound message processing
for a given tree via TestTreeProviderLite. It also adds the capability to
flush messages from a given tree. Together, these will allow finer
control over tests that use TestTreeProviderLite.
The test also changed the trees to be created with unique ids and as a
result some of the snapshots had to be updated.

This change is inspired by this comment
https://github.com/microsoft/FluidFramework/pull/24093/files#r2003975305.
There are scenarios where a test wants to carefully control which client
(tree) sequences or processes its ops and when while other clients
don't.


[AB#34266](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/34266)