IntervalCollection: Remove old value type infrastructure (#24174)
Interval collections was original build on an extension of maps called
value types, where map keys where like sub-dds. This hasn't been the
case for a very long time, but many of the types used for that long gone
scheme was still floating around in the interval collection code.

This change simplifies the code by removing those outdated abstractions,
and inlining or streamlining the necessary implementations. This
unraveling of outdate abstractions makes the code easier to maintain and
modify, while not changing any functionality.