(tree) Optimize op bunch processing via branch merge instead of individual commit rebases (#24093)
## Description
This PR adds second set of optimizations to shared tree with the op
bunching feature. The first optimization was added in
https://github.com/microsoft/FluidFramework/pull/23732.

This optimization targets the scenarios where a bunch of ops arrive from
a remote client at ref seq# N and there are changes in the trunk at ref
seq# N+. Basically, the remote client didn't know about these changes
when it sent the ops. Currently, every remote change is rebased over
every change in the trunk making it an expensive operation because
rebases are expensive. This becomes increasingly expensive as the size
of the ops grow.
With this change, the remote changes are applied to the peer local
branch maintained by the client and then this branch is merged into the
trunk. Merge has an optimization where the rebasing of later commits in
the branch reuses the result of some computations (specifically
compositions) involved in rebasing earlier commits in the branch.

For example, let's say there is one change in trunk T1 and two remote
changes R1 and R2 come in.
- Without this change, there are 4 rebases:
  - R1 is rebased over T1 generating R1' which is added to the trunk.
  - R2 is rebased over inverse of R1.
  - The result of the above is rebased over T1.
- The result of the above is rebased over R1' generating R2' which is
added to the trunk.
- With this change, there are 2 rebases.
  - R1 is rebased over T1 generating R1' which is added to the trunk.
  - The inverse of R1, T1, and R1' are composed
- R2 is rebased over that composition, generating R2' which is added to
the trunk.
- Before the change: O((R + T) * R) is the number of rebases. After the
change: O(R) in the number of rebases + O(R + T) in the number of
compositions.
  
As the number of changes grow, so do the number of rebases with the
current logic. With the new logic, this goes down substantially.

## Performance improvement
There is an asymptomatic performance improvement which increases with the
number of inbound ops and the number of ops in the trunk. With
100 inbound commits in the same bunch and 100 commits in the trunk, the
time to process goes from 860 ms second to 27 ms with a 97% performance
improvement.