fix(merge-tree): Fix partial lengths for reconnecting obliterate perspectives (#24193)
## Description

We did not correctly account for several partial lengths cases that can
happen with reconnecting perspectives when obliterates are possible.
This is mostly because unlike set removes, slice removes (obliterates)
make it possible for a client to remove a segment it had not yet seen,
which was not totally accounted for. Given recent changes in the area
and historical knowledge, it's also probable that some of the
perspectives we got wrong were codepaths we happened to not exercise
before (see e.g. recent changes to partial lengths which make sure that
an adjustment causing a segment to be seen as removed can only apply
when a corresponding adjustment causing the segment to have been seen as
inserted at all has been applied as well)

Co-authored-by: Abram Sanderson <absander@microsoft.com>