telemetry(container-runtime): add telemetry for "attached" subsequent first "saved" event (#24075)
## Description
Fixes
[AB#31457](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/31457)

This change adds telemetry event for when the container runtime attach
state is officially "Attached" and the first "saved" event that is
emitted immediately following attachment. This change is in response to
this
[repair-item](https://dev.azure.com/fluidframework/internal/_workitems/edit/31457):

"Loop team reported delay in receiving "saved" event after container
attach:
https://portal.microsofticm.com/imp/v5/incidents/details/588496721/summary.
At the time of creating this repair item, the root-cause is unknown.
However, there seems to be the need to having some telemetry around when
the "saved" event gets triggered after container is first attached, so
that we know are able to analyze the validity of container's state. Can
consider logging `isDirty` and other relevant properties as well."

Filling this gap in telemetry will help this current investigation as
well as future issues that come up in this area.