Revert "Update development and CI to Node 20" (#24209)
Reverts microsoft/FluidFramework#21965

Looks like the `test:realsvc:local:report:full` seems to run out of JS
heap space occasionally on CI with updated Node. Also sometimes a single
test times out.

As this is nondeterministic its taken a while to pin down but it seems
like this update is the cause.