(compat) Added support for layer compat validation at the Runtime / DataStore boundary (#24166)
## Description

Added logic to support layer compatibility validation at the Runtime /
DataStore layer boundary.

For context, please look at #22877 which added the same support to the
Runtime / Loader boundary. This logic added by this PR is the same as in
#22877.

### Supported features
- Both the Runtime and DataStore layers have a `supportedFeatures`
property that includes a set of features that a layer supports. This is
advertised to the other layer at the layer boundary.
- Both the layers have a `requiredFeatures` array that includes a set of
features that it requires the other layer to support in order to be
compatible. Note that this is internal to that layer and is not
advertised.
- At runtime, each layer validates that all the entries in
`requiredFeatures` is present in the `supportedFeatures` of the other
layer. If not, it throws a usage error that says layers are not
compatible.
- Any change that adds new behavior or changes existing behavior across
a layer boundary must add an entry to the supported features set. For
example, changes such as "change the format of the the summary returned
by the DataStore to Runtime during summarize call" should add an entry
to the supported features set.

### Generation
- In addition to `supported features`, a `generation` is also added to
both the layers. This is advertised to the other layer at the layer
boundary.
- `Generation` is an integer which will be incremented every month. It
can be used to validate compatibility more strictly (time-based) than
`supported features` where layers are incompatible only if there are
unsupported features.
_Note: The logic to update the generation will be added in a later PR.
See
[AB#27054](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/27054)
for a proposed solution._
- One key reason for adding generation is to have a clear cut off point
where we officially stop supporting layer combinations. Our compat tests
will test combinations up to this cut off point. The idea is that we
test what we support and whatever we don't test should not be supported.
This will help us achieve that.
- Both the layers have a `minSupportedGeneration` that it needs the
other layer to be at. Note that this is internal to that layer and is
not advertised.
- At runtime, each layer validates that the `generation` of the other
layer is `>= minSupportedGeneration`.
- For example, say that the Runtime is at generation 10 and has a
minSupportedGeneration of 7 for the DataStore layer (3 month support
window). If it sees that DataStore's generation is lower than 7, the
layers are incompatible.

## Backwards compatibility
To support older layers before layer compat enforcement is introduced,
the `minSupportedGeneration` is set to 0 whereas `generation` starts at
1. For older layers that did not implement `ILayerCompatDetails`, their
generation is set to 0 and `supported features` is set to an empty set.
So, their `generation` is compatible until we update the
`minSupportedGeneration`.


[AB#20806](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/20806)