Workaround TypeScript type checking optimization which has undesirable impacts on assignability in edge case (#24234)
## Description

TypeScript's type checker has some optimizations which impact the
assignability of types. See
https://github.com/microsoft/TypeScript/issues/59049

In this case the optimization is preventing non-empty POJO emulation
mode object schema from being assignable to `ObjectNodeSchema`.

This change introduces a workaround for this type checking behavior.

This workaround is not applied to the `objectRecursive` API as it does
not support POJO mode use and is expressed in a way where this issue
likely wouldn't occur anyway.