Added support to cache getSession responses (#24236)
## Description

We have seen a large number of `getSession` API calls for the same
documentId. This PR introduces a cache to store the responses for a
short duration of 5 mins to reduce the number of DB calls. It also
caches deleted documents for a longer time of 24 hours to prevent DB
calls. It also caches deleted docs info in the data-plane delete API.

For EC, since documents are limited to the cluster Redis, we set
`getSession` cache TTL as the TTL of the EC. THis would reduce the DB
access for an EC to almost 0.

All changes were tested locally using Docker.