Made cache TTLs configurable and made caching cleaner (#24243)
## Description

This PR makes cache TTLs configurable for deleted containers and also
for DC's that are not deleted. The default values for these are 20 mins
and 5 mins respectively.

It also cleans up the caching for deleted sessions. Instead of using the
cache value as `404`, we now use an object of type `ISession` but set
all urls to `""`. For these sessions, we set `isSessionAlive` and
`isSessionActive` to false as well.