server: upgrade to Node 20 (#24240)
## Description

Following #24215, this upgrades the server Dockerfiles to use Node 20.
Only thing that broke were some Sinon Fake Timers that weren't applying
to the common-utils package's performance export. Not certain why, but
addressed with a quick mock.

The chosen Node Docker image has 1 high-severity vulnerability for the
npm cross-spawn package. However, this is not a regression from v18, and
it has 1 fewer high-sev linux vulns, making this a net reduction in
vulnerability.