Update handlecache.ts (#24233)
## Description

While testing the performance of SharedMatrix when loading large CSV
datasets of 100k-500k rows and 10-12 columns, it was identified that
cacheMiss() contributes significantly to the overall performance. This
change removes the spread operator and reduces the number of copies and
edits made to the handles array. Performance testing shows a 3x
improvement in time taken to load large dataset, and a 50-95% reduction
in memory usage.