test(merge-tree): Remove ReconnectTestHelper (#24242)
## Description

This unifies `ReconnectTestHelper` and `PartialSyncTestHelper` and
renames it to `ClientTestHelper`. I've changed the way that resubmit
flows work a bit in doing so in a way which is hopefully more intuitive
to use (see documentation on methods: that should be improved if it is
not), and have some further ideas that I want to push to a different PR
which I think would get us even closer to the ideal of minimizing the
class of 'test-only code that only exists for ergonomics'.

It was mostly obliterate tests that used these helpers. While
refactoring, I've also adjusted our obliterate suites so they run with
and without `mergeTreeEnableSidedObliterate`, as the code paths do have
small differences before they converge (non-sided obliterates normalize
to a sided representation).

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>