feat(tree): Add (prototype/internal) table schema APIs (#24211)
Adds prototype Table schema factory APIs. These are exported as
`@internal` for now so we can start evaluating them within the repo.
They will be promoted to `@alpha` in a future PR once we have done some
basic evaluation.

See unit tests for examples of how the APIs are used.

Note for reviewers: I am still working on filling in missing unit tests.
But the APIs themselves are, I think, in a place where they are ready
for review.