Logging socket.io HTTP requests and HTTP upgrade requests for socket.io (#24250)
## Description

Socket.IO uses HTTP requests to establish a WebSocket connection with
the server. Here's how it works:

Initial HTTP Request (Handshake):

The client first sends an HTTP request to initiate a "handshake" with
the server. This allows the server and client to exchange information
and establish the connection.

By default, this is an HTTP GET request to the server’s Socket.IO
endpoint (e.g., /socket.io/?EIO=4...).

WebSocket Upgrade Request:

After the handshake, if the server and client both support WebSockets,
the HTTP connection is upgraded to a WebSocket connection using an HTTP
Upgrade request. This enables bi-directional, low-latency communication.

In this PR, I am adding some telemetry to track socket.io requests and
upgrade requests to the HTTP server.