Encode handles in ContainerRuntime (#24230)
When submitting ops up through the ContainerRuntime, as soon as a handle
in the op is encoded, we can no longer access the target object
directly. This lossy change is only needed ahead of JSON.stringify, so
it's better to delay it until as late as possible.

In particular, to support upcoming "Staging Mode", we will need to
preserve the original (non-encoded) ops such that resubmit and rollback
get the original handles. This is a prerequisite to that change (at this point,
resubmit and rollback still get a copy roundtripped through string)

---------

Co-authored-by: Tony Murphy <anthony.murphy@microsoft.com>