Add TestFluidObjectInternal that avoids using SharedMap (#23949)
## Description

This change avoids impacting the alpha+legacy APIs around test objects,
and just slips in a way to use an internal one which does not use
SharedMap. This is then used in tree tests so they don't create needless
shared maps.

TestFluidObject and ITestFluidObject are the only use
@fluidframework/map in test-utils, so if it gets deprecated and removed,
and replaced by the new TestFluidObjectInternal this change adds, the
dependency on @fluidframework/map could be removed, which may be enough
to allow test-utils to be used to test map like it is for other DDSes.