Switch to using DocumentSchema to enable blob placeholders (#24262)
DocumentSchema is a more appropriate tool than a feature flag, because
once it is enabled for a document it permanently puts older incompatible
clients at risk of encountering op patterns they aren't prepared for
(namely, an op containing the handle before the blob attach op).

This PR primarily switches to that mechanism, but it also revealed that
I had a bug in my flag enablement in blobManager.spec.ts that was
causing the tests to run only in legacy mode (I wasn't paying attention
to how/when injectedSettings got wiped out). After fixing that test bug,
the now-functioning tests revealed a couple other small bugs:

1. The pendingBlobs created by the placeholder handles should be marked
attached immediately upon creation, because they are created by the act
of attaching the handle.
2. waitForBlob in the test had an inverted condition - it should be
waiting to see its blob appear in the unprocessedBlobs (exists in main
but must not be causing problems there)
3. An 0x38f assert in one of the tests that I haven't debugged fully
yet, though it may just be a test bug due to the handle management.

Because blobManager.spec.ts expects to have such intimate knowledge of
the BlobManager internals, it might be reasonable to consider a broader
rework of its test coverage for placeholder blobs (since it subverts the
expectations about handle and upload behavior). Maybe for later
consideration.


[AB#34418](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/34418)

[AB#34459](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/34459)