ci: Fix the fix on syntax in yaml file (#24291)
## Description

Follow up to https://github.com/microsoft/FluidFramework/pull/24290 with
further syntax fixing. `${{}}` syntax gets coalesced to a blank _at
compile time_ when it produces no output, so that resulted in a syntax
error. Wrapping the use of that syntax with quotes should result in an
empty string which should have the intended effect.

Using
https://github.com/microsoft/FluidFramework/blob/main/tools/pipelines/templates/build-npm-client-package.yml#L677
as reference that this should work.