feat(merge-tree): Support resubmission of obliterates (#24253)
## Description

This updates merge-tree's `resubmit` to support resubmission for
obliterates. It takes generally the same strategy that we have used in
`intervalCollection`, with a few subtleties around updating additional
bookkeeping we have in `ObliterateInfo` and on segments.

As part of this PR, I've added some unit tests to the obliterate
reconnect file and turned on reconnection in the obliterate fuzz test
suite.

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>