Streamline logic around grouping/compressing in outbox.flushInternal (#24288)
The cases around when to group and/or compress when flushing a batch are
nuanced. This step is called "Op Virtualization" since it takes a batch
of raw local runtime ops and results in a completely different set of
ops to transmit the data to the server.

There is a pretty simple decision tree to follow, now that compression
only works on singleton batches:

1. If grouping is enabled, try to group (it's a no-op if already a
single message)
2. If we are left with a multi-message batch, don't compress (this is no
longer supported), just return.
3. If we can / need to compress, do it

I also updated `OpCompressor.compressBatch` to take a singleton batch
(enforced by the type system).