feat(ContainerRuntime): Flush on DeltaManager's "op" event (back-compat for old loader) (#24276)
We need to flush whenever incoming ops are processed (even non-runtime
ops), since we track the reference sequence number of the in-progress
batch and throw if it ever changes mid-batch. For this reason, we
recently started calling `flush` in `ContainerRuntime.process`.

In old loader code, `ContainerRuntime.process` is only called for
runtime messages, so it is possible for the reference sequence number to
advance without ContainerRuntime flushing.

This change adds another flush call in response to DeltaManager's "op" event
to close this gap.

It's ok to call flush extra, it's a cheap no-op if there's nothing
there.