(tree) Fix for recursion not detected in unhydrated nodes (#24246)
## Bug
Recursively assinging an unhydrated and unparented node to itself hangs
forever. Note that doing the same for a parented node does throw a usage
error (added a test for this scenario)

## Fix
The root-cause of this bug is that it is stuck in a while loop in
`TreeNodeKernel` trying to emit `subtreeChangedAfterBatch` event for the
sub-tree of the node.
Added logic in `adoptBy` to detect recursion and throw a usage error.
This is where the logic to detect this for parent node happens.


[AB#32207](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/32207)