refactor(container-runtime): Separate types LocalBatchMessage v. OutboundBatchMessage in Outbox and related code (#24287)
This is a glorified "rename", splitting the type
`BatchMessage` into two cases: `LocalBatchMessage` and
`OutboundBatchMessage`, converting between the two when necessary.

The intention is to change no behavior with this change. Here are the
non-typing changes (that affect runtime code):

* Renaming the property `contents` to `serializedOp` for the Local batch
case
* Shallow-copy the whole batch and each message in
`Outbox.virtualizeBatch` to convert from `LocalBatch` to
`OutboundBatch`.
* This is where we copy `serializedOp` over to `contents`, which paves
the way for `serializedOp` to stop being serialized in the next PR,
which is the end goal here
* Minor refactoring of `createEmptyGroupedBatch` to support both
Local/Outbound forms

### Context

As ops are accumulated into a batch in the ContainerRuntime, they're
currently serialized immediately. However, we plan to keep those
unserialized, and only stringify when virtualizing (grouping,
compressing) in preparation to send to the server. See #24281.

This is the first step there. We should model pre/post virtualized
batches differently because they are different. pre-virtualization batch
messages are called `LocalBatchMessage` and post-virtualization are
called `OutboundBatchMessage`.