build(build-tools,client): Remove lerna.json files (#24273)
Removes lerna.json files since they are no longer used. Also removes
pipeline references to lerna.json, and cleans up a few other references
in other files.

There is a reference to lerna in a comment in the mocha-test-setup
project. I left it as-is, because it still seems relevant.

Split off from #24072.