Added a document denyList in Alfred and Nexus (#24298)
## Description

This PR adds a document denyList for Alfred and Nexus. This is similar
to what we have in Historian. This is added to the all Alfred APIs. The
list has two features:

1) A list of tenants to block: This list can allow blocking of all
requests from a given tenant. Useful for situations when we get DDoS-ed
by 1 tenant.
2) A list of documents to block: Blocks documents that are large in size
and can cause OOM issues. Once change here from the previous
implementation is that now we no longer use the tenantId -> documentId
map. This is to make OCE life easier. Once GUID is easier to
type/copy-paste than two GUIDs. Since DocIds are GUIDS chances of
collision are negligible.

For some `/documents` APIs, we can skip document level checks as these
ops are either `GET` for document data and not ops or are `DELETE` calls
to delete the document.

This PR also has TODOs to remove the implementation in Historian and use
the implementation in routerlicious instead.