Reduce use of symbols where no longer needed (#24302)
## Description

Back when the lazy entity types where originally implemented, the
FlexTree API for object nodes exposed fields as properties, which could
collide with members used in the implementation, so symbols were used to
avoid this.

As this these properties are no longer used, these APIs can be migrated
away from symbols for cleaner code and easier debugging.