Fix exception thrown when debugging tests using test browser in tree (#24304)
## Description

When using the vs-code test browser as configured by the tree workspace
to debug tests, a handled exception is thrown who's messages suggested
this change. Making this change seems to work, and the exception is gone
which makes debugging with break on caught exceptions nicer.

## Breaking Changes

This changes when the debugger breaks, making it break less ;)