Delay load parts of summary module to reduce the container runtime bundle size. (#24161)
## Description

This redccues bundle size of container runtime by ~20KB.
Delay load parts of summary module like RunningSummarizer,
SummaryGenerator etc to reduce the container runtime bundle size. This
effort is done as part of reducing fluid bundle size for loop so as to
make uop for some of the increased bundle size due to new shared tree.

---------

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>