Update Uuid package from 9 to 11 (#24321)
## Description

Update the version of the uuid package. It is now written in TypeScript,
avoids throwing a caught exception on load and now officially supports
node 20 (which we use).

More details in [the
changelog](https://github.com/uuidjs/uuid/blob/main/CHANGELOG.md).

The updated version of uuid, at least the subset of it pulled into the
tree package is 1041 bytes, where the old one was 868, so this does
regress bundle size by 173 bytes uncompressed: the regression should be
less for the compressed and parsed sizes.

I think 173 bytes is acceptable for the developer workflow improvement
of not having exceptions on startup (once the server update is pulled
into client) or dealing with the separate types package. The tree bundle
being considered here is 357568 making this a regression of
approximately 0.048%, about the same as we would see from adding a nice
UssageError to an API.