fix(merge-tree): Apply minor fixes to obliterate in re-resubmit grouped batching scenario (#24292)
## Description

With recent fixes to merge-tree, we're able to enable a bunch of
previously skipped fuzz+regression tests around grouped batching. This
PR does so. I also checked the status of turning on reconnect + grouped
batching for obliterate and interval collection. There was one minor /
easily fixed issue which this PR addresses and adds regression tests
for. The remaining issues look like they occur much more rarely and are
probably fundamentally related to known bugs in the segment
normalization codepath. I've updated references in various comments to
reflect this.

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>