fix(container-runtime): Always close the container if ContainerRuntime.flush throws (#24318)
Flush is often called from secondary callstacks such that an error
thrown will not propagate up to user code. However, failure to flush is
a fatal error that we cannot recover from.

This change adds a try-catch to `ContainerRuntime.flush` to ensure that
the container is always closed if an error is thrown. I also went
through all the errors thrown underneath there and made some tweaks to
make them easier to interpret. e.g. using `DataProcessingError` more,
since this is an error type we expect consumers to monitor closely.