fix(check:policy): Support running checks from the routerlicious folder (#24249)
Policy check's handling of paths was inconsistent, which made it
impossible to run policy check anywhere outside the root of the repo,
like in the routerlicious folder.

The ultimate problem is that some handlers need absolute paths, some
need repo-relative, and some need paths to be adjusted based on the
current working directory. Prior to this change, handlers and resolvers
were passed repo-relative paths, which they would use as-is. This works
fine when running from the root, since the paths will always resolve.
But from other paths in the repo, those repo-relative paths did not
resolve.

**Handlers and resolvers now always receive absolute paths to both the
git repo root and the file being checked.** From that, they can derive
relative paths to the files if needed. This was the least amount of
churn since many handlers just read files and write back to them, so by
passing absolute paths, those handlers remained unchanged. Only handlers
that themselves deal with paths needed to be updated.

I did _not_ go through each handler and review it. I only made changes
that were needed to support running checks from the routerlicious
folder, and confirmed that running from the root and routerlicious
folder both functioned as expected.

I'm not sure a more robust approach is needed. My goal is only to
unblock #24072, and this does that. I think additional bugs can be fixed
if they come up.

### Other changes

Also removes the long-deprecated `--exclusions` flag.