Re-enable: Shortids for Datastores and DDSes (#24231)
This PR enables use of short ids by default. Along with having a
kill-switch configuration available to turn the feature off if needed).
The PR also fixes the corresponding test. The was implemented with
`assetInvert` to negatively test for special characters and short ids
when the bug was not fixed.

Note:
This PR is a follow up from a sequence of fixes that have went in
before, to ensure cross client and cross layer compat.
Short summary - short ids previously generated characters that changed
their form when any code used `encodeURIComponent()`. That has been
fixed by - 1 updating the shortids generator code to only use encode
safe characters + removal of `encodeURIComponent` where it was being
used.
https://github.com/microsoft/FluidFramework/pull/21680
https://github.com/microsoft/FluidFramework/pull/24097
https://github.com/microsoft/FluidFramework/pull/24032


Fixes
[AB#8568](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8568)