refactor(container-runtime): Stop estimating batch size when accumulating a local batch  (#24309)
We are preparing to update `LocalBatchMessage` to hold the original
runtime op, not the serialized op. So it will no longer be practical to
estimate the content size of each op (we don't want to do an extra JSON
stringify just for this purpose).

The downside is if compression is disabled and someone submits an op
that will likely push the batch over the threshold, they won't get that
error under that submit callstack like they used to. Instead it will
only surface when the container closes during flush.

Here's a comparison of the places we would fail before / after this
change:

Before:
* If no compression, then if estimsated socket size is too big, throw when
submitting. (in `Outbox.addMessageToBatchManager`)
* "If no compression" because `hardLimit` would be infinity if
compression was enabled
* If yes compression (and it happens), then throw if it was insufficient
(in `Outbox.virtualizeBatch`)

Now:
* If yes compression (and it happens), then throw if it was insufficient
(in `Outbox.virtualizeBatch`) -- _unchanged from before_
* Otherwise (no compression happened, or it did and appeared
sufficient), if est socket size is too big, throw before send (in
`Outbox.sendBatch`)
  * This case used to be possible even after compression if there were
soooo many empty placeholders that it pushed it over the limit. Empty placeholders aren't a thing anymore.
  * But now, if compression is sufficient or unnecessary, this will surely
pass as well. So this is just guarding the case where compression was
disabled, so it's quite equivalent to the original check during submit.